# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GlobeTrek Overseas Student Application & Payment API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  _id: string;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  success: boolean;
  message?: string;
  token: string;
  user: User;
}

export interface ProfileResponse {
  success: boolean;
  user: User;
}

export interface University {
  _id: string;
  name: string;
  country: string;
  course: string;
  applicationFee: number;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface UniversityInput {
  name: string;
  country: string;
  course: string;
  /** @minimum 0 */
  applicationFee: number;
  description?: string;
}

export interface UniversityListResponse {
  success: boolean;
  count?: number;
  universities: University[];
}

export interface UniversityResponse {
  success: boolean;
  university: University;
}

export interface ApplicationInput {
  universityId: string;
  statement?: string;
}

export type ApplicationStatus = typeof ApplicationStatus[keyof typeof ApplicationStatus];


export const ApplicationStatus = {
  pending: 'pending',
  under_review: 'under_review',
  accepted: 'accepted',
  rejected: 'rejected',
} as const;

export type ApplicationPaymentStatus = typeof ApplicationPaymentStatus[keyof typeof ApplicationPaymentStatus];


export const ApplicationPaymentStatus = {
  unpaid: 'unpaid',
  paid: 'paid',
} as const;

export interface Application {
  _id: string;
  student?: User;
  university?: University;
  status: ApplicationStatus;
  /** @nullable */
  statement?: string | null;
  paymentStatus: ApplicationPaymentStatus;
  createdAt: string;
}

export interface ApplicationResponse {
  success: boolean;
  application: Application;
}

export interface ApplicationListResponse {
  success: boolean;
  count?: number;
  applications: Application[];
}

export interface PaymentOrderInput {
  applicationId: string;
}

export interface PaymentOrderResponse {
  success: boolean;
  message?: string;
  clientSecret: string;
  paymentIntentId: string;
  amount: number;
  currency: string;
}

export interface PaymentVerifyInput {
  paymentIntentId: string;
}

export type PaymentStatus = typeof PaymentStatus[keyof typeof PaymentStatus];


export const PaymentStatus = {
  created: 'created',
  succeeded: 'succeeded',
  failed: 'failed',
} as const;

export interface Payment {
  _id: string;
  stripePaymentIntentId: string;
  amount: number;
  currency: string;
  status: PaymentStatus;
  application?: Application;
  createdAt: string;
}

export interface PaymentVerifyResponse {
  success: boolean;
  message?: string;
  payment?: Payment;
}

export interface PaymentHistoryResponse {
  success: boolean;
  count?: number;
  payments: Payment[];
}