# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GlobeTrek Overseas Student Application & Payment API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: universities
    description: University management
  - name: applications
    description: Student applications
  - name: payments
    description: Payments

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already registered

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login and get JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials

  /auth/profile:
    get:
      operationId: getProfile
      tags: [auth]
      summary: Get current student profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"

  /universities:
    get:
      operationId: listUniversities
      tags: [universities]
      summary: List all universities
      responses:
        "200":
          description: List of universities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UniversityListResponse"
    post:
      operationId: createUniversity
      tags: [universities]
      summary: Create a university
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UniversityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UniversityResponse"

  /universities/{id}:
    put:
      operationId: updateUniversity
      tags: [universities]
      summary: Update a university
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UniversityInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UniversityResponse"
    delete:
      operationId: deleteUniversity
      tags: [universities]
      summary: Delete a university
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted

  /applications:
    post:
      operationId: createApplication
      tags: [applications]
      summary: Apply to a university
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResponse"
        "409":
          description: Already applied
    get:
      operationId: listApplications
      tags: [applications]
      summary: List my applications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Applications list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationListResponse"

  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get one application
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResponse"

  /payments/create-order:
    post:
      operationId: createPaymentOrder
      tags: [payments]
      summary: Create Stripe PaymentIntent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentOrderInput"
      responses:
        "201":
          description: Payment order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentOrderResponse"

  /payments/verify:
    post:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify and confirm payment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentVerifyInput"
      responses:
        "200":
          description: Payment verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVerifyResponse"

  /payments/history:
    get:
      operationId: getPaymentHistory
      tags: [payments]
      summary: Get payment history
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Payment history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentHistoryResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        phone:
          type: string

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    User:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [_id, name, email, createdAt]

    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [success, token, user]

    ProfileResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: "#/components/schemas/User"
      required: [success, user]

    University:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        country:
          type: string
        course:
          type: string
        applicationFee:
          type: number
        description:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [_id, name, country, course, applicationFee, createdAt]

    UniversityInput:
      type: object
      required: [name, country, course, applicationFee]
      properties:
        name:
          type: string
        country:
          type: string
        course:
          type: string
        applicationFee:
          type: number
          minimum: 0
        description:
          type: string

    UniversityListResponse:
      type: object
      properties:
        success:
          type: boolean
        count:
          type: integer
        universities:
          type: array
          items:
            $ref: "#/components/schemas/University"
      required: [success, universities]

    UniversityResponse:
      type: object
      properties:
        success:
          type: boolean
        university:
          $ref: "#/components/schemas/University"
      required: [success, university]

    ApplicationInput:
      type: object
      required: [universityId]
      properties:
        universityId:
          type: string
        statement:
          type: string

    Application:
      type: object
      properties:
        _id:
          type: string
        student:
          $ref: "#/components/schemas/User"
        university:
          $ref: "#/components/schemas/University"
        status:
          type: string
          enum: [pending, under_review, accepted, rejected]
        statement:
          type: ["string", "null"]
        paymentStatus:
          type: string
          enum: [unpaid, paid]
        createdAt:
          type: string
      required: [_id, status, paymentStatus, createdAt]

    ApplicationResponse:
      type: object
      properties:
        success:
          type: boolean
        application:
          $ref: "#/components/schemas/Application"
      required: [success, application]

    ApplicationListResponse:
      type: object
      properties:
        success:
          type: boolean
        count:
          type: integer
        applications:
          type: array
          items:
            $ref: "#/components/schemas/Application"
      required: [success, applications]

    PaymentOrderInput:
      type: object
      required: [applicationId]
      properties:
        applicationId:
          type: string

    PaymentOrderResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        clientSecret:
          type: string
        paymentIntentId:
          type: string
        amount:
          type: integer
        currency:
          type: string
      required: [success, clientSecret, paymentIntentId, amount, currency]

    PaymentVerifyInput:
      type: object
      required: [paymentIntentId]
      properties:
        paymentIntentId:
          type: string

    Payment:
      type: object
      properties:
        _id:
          type: string
        stripePaymentIntentId:
          type: string
        amount:
          type: integer
        currency:
          type: string
        status:
          type: string
          enum: [created, succeeded, failed]
        application:
          $ref: "#/components/schemas/Application"
        createdAt:
          type: string
      required: [_id, stripePaymentIntentId, amount, currency, status, createdAt]

    PaymentVerifyResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        payment:
          $ref: "#/components/schemas/Payment"
      required: [success]

    PaymentHistoryResponse:
      type: object
      properties:
        success:
          type: boolean
        count:
          type: integer
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
      required: [success, payments]