# lib/api-zod/src/generated/types/applicationListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GlobeTrek Overseas Student Application & Payment API
 * OpenAPI spec version: 0.1.0
 */
import type { Application } from './application';

export interface ApplicationListResponse {
  success: boolean;
  count?: number;
  applications: Application[];
}