# lib/api-zod/src/generated/types/profileResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GlobeTrek Overseas Student Application & Payment API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface ProfileResponse {
  success: boolean;
  user: User;
}