# lib/api-zod/src/generated/types/universityInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GlobeTrek Overseas Student Application & Payment API
 * OpenAPI spec version: 0.1.0
 */

export interface UniversityInput {
  name: string;
  country: string;
  course: string;
  /** @minimum 0 */
  applicationFee: number;
  description?: string;
}